royality_bonus

+++ database/migrations/2021_10_30_120928_create_cash_wallets_table.php
@@ -16,7 +16,7 @@ class CreateCashWalletsTable extends Migration
         Schema::create('cash_wallets', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id');
-            $table->integer('bonus_amount');
+            $table->float('bonus_amount');
             $table->string('method')->nullable();
             $table->text('note')->nullable();
             $table->timestamps();
